Midterm Project
All page production must be generated using JavaScript. The only HTML
code permitted must be encased as part of a JavaScript statement. All
styling must be accomplished with CSS. In-line or an external CSS
document is acceptable.

Your mid-term projects are:

1. Create a 2 x 2 table of images that relate to the spring season. All
images are to be sized identically and any image will fit into its own
cell. Whether the table has a border or no border is your decision.

The images must be array-based. All even number images will also link
to a garden nursery in the Boston area. The odd images will connect to
a flower shop in the Boston area.  The page is to have a header that
reads "Welcome to Spring".  The page is to have a complimentary
background.

2. You are to provide a conversion page that will convert:

a. Dollars to Euros based on and exchange rate of `.93 Euros per dollar.
b. Dollars to Francs on an exchange rate of 1.035 Francs per dollar
c. Dollars to pesos on an exchange rate of 15.63 pesos per dollar

This is to be in a table format with table headings indicating which
conversion is being presented. The table colors are to alertnate. The
table values are to be between 1 and 100. The loop increment value will
be requested from the viewer. This means that your table may list 1-100
or 1,3,5,7,9 or 5,10,15,20, etc.

3. You will produce an on line menu offering for a hotel that will
display a breakfast menu, a lunch menu, or a dinner menu, and a night
menu based on the time of day. This page must work for all time
periods. The menu must be a complete offering for that time period.
The page must be styled.

4. Produce a sorted list of array-based buttons that will connect to 10
ski resorts in the New England area (okay if you go outside of New
England) . The buttons should have a rounded corner transition effect
when they are hovered.  The page must be styled with an appropriate
header.